+++ test_lib/id_jungle.ex
@@ -5,6 +5,5 @@ defmodule IDJungle do
     nullable: true,
     format: :uuid,
     pattern: ~r/[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}/,
-    example: "193202fc-ab55-4824-8ec8-5bef1201d9eb",
-    struct?: false
+    example: "193202fc-ab55-4824-8ec8-5bef1201d9eb"
 end